feat: adds onboarding
