+++ src/cli/onboarding/types.ts
@@ -1,0 +1,4 @@
+export interface OnboardingSection {
+  title: string;
+  lines: readonly string[];
+}
