+++ tests/e2e/scripts/run-e2e-tests.sh
@@ -424,6 +424,11 @@ echo "=============================="
 /app/scripts/test-agent-dev-setup.sh
 print_result $? "Agent and local dev setup tests completed"
 
+echo "\nRunning Onboarding Tests..."
+echo "=============================="
+/app/scripts/test-onboarding.sh
+print_result $? "Onboarding tests completed"
+
 echo "\n📝 Running RC File Suggestion Tests..."
 echo "=============================="
 /app/scripts/test-rc-file-suggestion.sh
